+++ src/@core/components/Header.tsx
@@ -7,7 +7,8 @@ export default function Header() {
         <div className="flex items-center justify-between p-6 border-b-2 border-black">
           <a href="/">
             <img src={logo} width={300} />
-            <span className="text-[#396edf] font-['Tenada'] italic ">
+            {/* 로고 임시로 숨김 */}
+            <span className="text-[#396edf] italic font-size text-5xl font-black hidden">
               ONECLICK
             </span>
           </a>
